API reference

HyperCalibration.Biaxial — Type

Biaxial kinematics allow to evaluate the stress response at one Gauss point.

HyperCalibration.CreepProtocol — Type

A sequence of times at a constant stretch.

HyperCalibration.CyclicLoadingProtocol — Type

A sequence of stretches at a constant rate and time step.

HyperCalibration.DielectricMeasurement — Type

The dielectric parmittivity recoded during a dielectric test.

HyperCalibration.DielectricSpectroscopyTest — Type

Experiment data with frequency-dielectric permittivity (BDS).

HyperCalibration.DifferentialScanningCalorimetryTest — Type

Experiment data with temperature-specific heat capacity (DSC).

HyperCalibration.ElectricalCondition — Type

An experiment at a constant voltage.

HyperCalibration.ExperimentData — Type

ExperimentData{M, P, C, G}

An experiment holds a measurement M that is a function of a Protocol P -defining the independent variables- and conditionc C -environtmental static variables- and geometry G -for the reference specimen configuration-. Additionally, a weight can be assigned to the experiment to define its importance in the calibration process.

HyperCalibration.FrequencySweepProtocol — Type

A sequence of frequencies.

HyperCalibration.IsothermalCondition — Type

An experiment at a constant temperature.

HyperCalibration.MechanicalTest — Type

Super type for ExperimentData with mechanical measurement (stress) and mechanical protocol (stretch).

HyperCalibration.PlateGeometry — Type

A specimen geometry with a reference thickness and cross-sectional area.

HyperCalibration.QuasiStaticProtocol — Type

A sequence of stretches.

HyperCalibration.SequentialProtocol — Type

A sequential protocol defines a series of sub-protocols to be executed in order. E.g. a relaxation test is defined by a fast loading stage (CyclicLoadingProtocol) followed by a long holding stage (CreepProtocol).

HyperCalibration.StandardCondition — Type

A standard condition with no additional parameters.

HyperCalibration.TemperatureSweepProtocol — Type

A sequence of temperatures at constant rate.

HyperCalibration.TensileMeasurement — Type

The nominal stress recoded during a tensile test.

HyperCalibration.ThermalMeasurement — Type

The specific heat capacity recoded during a thermal test.

HyperCalibration.ThermalTest — Type

Super type fo ExperimentData with thermal measurement (specific heat capacity) and thermal protocol (temperature).

HyperCalibration.ThermoDielectricTest — Type

Super type for ExperimentData with dielectric measurement (permittivity) and thermal protocol (temperature).

HyperCalibration.ThermoElectricalCondition — Type

An experiment at a constant temperature and voltage.

HyperCalibration.Uniaxial — Type

Uniaxial kinematics allow to evaluate the stress response at one Gauss point.

HyperCalibration.UniaxialCyclicLoadingTest — Type

Experiment data with constant rate uniaxial stretch-stress and standard conditions.

HyperCalibration.UniaxialQuasiStaticTest — Type

Experiment data with uniaxial quasi-static stretch-stress and standard conditions.

HyperCalibration.UniaxialRelaxationTest — Type

Experiment data with uniaxial relaxation stressess and standard conditions.

HyperCalibration.UniaxialThermalCyclicLoadingTest — Type

Experiment data with constant rate uniaxial stretch-stress and thermal conditions.

HyperCalibration.UniaxialThermalQuasiStaticTest — Type

Experiment data with uniaxial quasi-static stretch-stress and thermal conditions.

HyperCalibration.UniaxialThermalRelaxationTest — Type

Experiment data with uniaxial relaxation stressess and thermal conditions.

HyperCalibration.UniaxialThermoElectricCyclicLoadingTest — Type

Experiment data with constant rate uniaxial stretch-stress and coupled thermo-electrical conditions.

HyperCalibration.electric_field — Method

Return the electric field for a given condition and geometry.

HyperCalibration.evaluate_cv — Method

Evaluate the specific heat capacity predicted by a constitutive model under the given thermal protocol, conditions and geometry of an experimental test.

HyperCalibration.evaluate_epsilon — Method

Evaluate the dielectric permittivity predicted by a constitutive model under the given thermal protocol, conditions and geometry of an experimental test.

HyperCalibration.evaluate_stress — Method

Evaluate the stress predicted by a constitutive model under the given mechanical protocol, conditions, and geometry of an experimental test.

HyperCalibration.experiment_prediction — Method

Return the y_true and y_pred sampling points for a given constitutive model and experimental test.

HyperCalibration.frequencies — Method

Return the sequence of frequencies for a given protocol.

HyperCalibration.independent_variable — Method

Return the sequence of the independent variable for a given protocol.

HyperCalibration.loss — Method

Compute the loss function for a given constitutive model and one/multiple experimental tests as the weighted normalized squared difference between the predicted and the measured values.

HyperCalibration.max_stretch — Method

Return the maximum stretch for a givben protocol.

HyperCalibration.rate — Method

Return the rate of independent variable increment for a given protocol.

HyperCalibration.stretches — Method

Return the sequence of stretches for a given protocol.

HyperCalibration.temperature — Method

Return the room temperature for a given condition.

HyperCalibration.temperatures — Method

Return the sequence of temperatures for a given protocol.

HyperCalibration.thickness — Method

Return the thickness for a given geometry.

HyperCalibration.time_step — Method

Return the constant time step for a given protocol.

HyperCalibration.voltage — Method

Return the applied voltage for a given condition.